Separate location upload persistence from queue coordination #299
